+++ globalTeardown.js
@@ -1,0 +1,8 @@
+const Pyroscope = require('@pyroscope/nodejs');
+
+module.exports = (config) => {
+  if (process.env.CI) {
+    Pyroscope.stopCpuProfiling();
+    Pyroscope.stopHeapProfiling();
+  }
+};