+++ packages/pyroscope-panel-plugin/.eslintrc.js
@@ -1,9 +1,0 @@
-const path = require('path');
-
-module.exports = {
-  extends: [path.join(__dirname, '../../.eslintrc.js')],
-  ignorePatterns: ['.eslintrc.js'],
-  parserOptions: {
-    tsconfigRootDir: __dirname,
-  },
-};