+++ pkg/agent/upstream/remote/remote.go
@@ -35,6 +35,10 @@ type Remote struct {
 
 type RemoteConfig struct {
 	AuthToken              string
+	BasicAuthUser          string
+	BasicAuthPassword      string
+	TenantID               string
+	HTTPHeaders            map[string]string
 	UpstreamThreads        int
 	UpstreamAddress        string
 	UpstreamRequestTimeout time.Duration
@@ -112,7 +116,7 @@ func (r *Remote) uploadProfile(j *upstream.UploadJob) error {
 	q.Set("units", string(j.Units))
 	q.Set("aggregationType", string(j.AggregationType))
 
-	u.Path = path.Join(u.Path, "/ingest")
+	u.Path = path.Join(u.Path, "ingest")
 	u.RawQuery = q.Encode()
 
 	r.Logger.Debugf("uploading at %s", u.String())
@@ -125,6 +129,14 @@ func (r *Remote) uploadProfile(j *upstream.UploadJob) error {
 
 	if r.cfg.AuthToken != "" {
 		request.Header.Set("Authorization", "Bearer "+r.cfg.AuthToken)
+	} else if r.cfg.BasicAuthUser != "" && r.cfg.BasicAuthPassword != "" {
+		request.SetBasicAuth(r.cfg.BasicAuthUser, r.cfg.BasicAuthPassword)
+	}
+	if r.cfg.TenantID != "" {
+		request.Header.Set("X-Scope-OrgID", r.cfg.TenantID)
+	}
+	for k, v := range r.cfg.HTTPHeaders {
+		request.Header.Set(k, v)
 	}
 
 	// do the request and get the response