+++ pkg/convert/pprof/streaming/vt_profile.go
@@ -7,9 +7,10 @@ import (
 )
 
 const (
-	opFlagCountStructs = 1 << 0
-	opFlagParseStructs = 1 << 1
-	opFlagParseSamples = 1 << 2
+	opFlagCountStructs           = 1 << 0
+	opFlagParseStructs           = 1 << 1
+	opFlagParseSamples           = 1 << 2
+	opFlagParseSamplesWriteBatch = 1 << 3
 )
 
 // revive:disable-next-line:cognitive-complexity,cyclomatic necessary complexity
@@ -19,14 +20,15 @@ func (p *VTStreamingParser) UnmarshalVTProfile(dAtA []byte, opFlag uint64) error
 	countStructs := opFlag&opFlagCountStructs == opFlagCountStructs
 	parseStructs := opFlag&opFlagParseStructs == opFlagParseStructs
 	parseSamples := opFlag&opFlagParseSamples == opFlagParseSamples
+	parseSamplesWriteBatch := opFlag&opFlagParseSamplesWriteBatch == opFlagParseSamplesWriteBatch
 	strWriteIndex := 0
 	locationWriteIndex := 0
 	functionWriteIndex := 0
-	sampelTypeWriteIndex := 0
-	p.strings = p.strings[:cap(p.strings)]
-	p.locations = p.locations[:cap(p.locations)]
-	p.functions = p.functions[:cap(p.functions)]
-	p.sampleTypes = p.sampleTypes[:cap(p.sampleTypes)]
+	sampleTypeWriteIndex := 0
+	p.strings = p.strings[:p.nStrings]
+	p.locations = p.locations[:p.nLocations]
+	p.functions = p.functions[:p.nFunctions]
+	p.sampleTypes = p.sampleTypes[:p.nSampleTypes]
 	if countStructs {
 		p.period = 0
 		p.nStrings = 0
@@ -94,10 +96,10 @@ func (p *VTStreamingParser) UnmarshalVTProfile(dAtA []byte, opFlag uint64) error
 			}
 			if parseStructs {
 				//p.sampleTypes = append(p.sampleTypes, valueType{})
-				if err := p.sampleTypes[sampelTypeWriteIndex].UnmarshalVT(dAtA[iNdEx:postIndex]); err != nil {
+				if err := p.sampleTypes[sampleTypeWriteIndex].UnmarshalVT(dAtA[iNdEx:postIndex]); err != nil {
 					return err
 				}
-				sampelTypeWriteIndex++
+				sampleTypeWriteIndex++
 			}
 			iNdEx = postIndex
 		case 2:
@@ -131,6 +133,11 @@ func (p *VTStreamingParser) UnmarshalVTProfile(dAtA []byte, opFlag uint64) error
 			}
 			if parseSamples {
 				if err := p.parseSampleVT(dAtA[iNdEx:postIndex]); err != nil {
+					return err
+				}
+			}
+			if parseSamplesWriteBatch {
+				if err := p.parseSampleWB(dAtA[iNdEx:postIndex]); err != nil {
 					return err
 				}
 			}