+++ pkg/remotewrite/client.go
@@ -106,6 +106,9 @@ func (r *Client) ingestInputToRequest(in *ingestion.IngestInput) (*http.Request,
 	}
 
 	req, err := http.NewRequest("POST", r.config.Address+"/ingest", bytes.NewReader(b))
+	for k, v := range r.config.Headers {
+		req.Header.Set(k, v)
+	}
 	if err != nil {
 		return nil, err
 	}
@@ -145,5 +148,12 @@ func (r *Client) enhanceWithAuth(req *http.Request) {
 
 	if token != "" {
 		req.Header.Set("Authorization", "Bearer "+token)
+	}
+
+	if r.config.BasicAuthUser != "" && r.config.BasicAuthPassword != "" {
+		req.SetBasicAuth(r.config.BasicAuthUser, r.config.BasicAuthPassword)
+	}
+	if r.config.TenantID != "" {
+		req.Header.Set("X-Scope-OrgID", r.config.TenantID)
 	}
 }