+++ webapp/javascript/components/AppSelector/Label.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export function Label() {
+  return <>Application:&nbsp;</>;
+}
+
+export const LabelString = 'Select Application';